+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [15.0.0] - 2024-08-16
+- Make spelling consistent for wallets
+
+## [14.0.0] - 2024-07-16
+
+## Added
+- Added V5 wallet support (thx Tonkeeper Team)
+- Fixed stack serialization for TonCenter v2 client (thx @aspite)
+
+## Fixed 
+- Types for different wallet versions
+
 ## [13.11.2] - 2024-05-31
 
 ## Added